+++ packages/devtools_app/test/inspector_v2/inspector_screen_test.dart
@@ -407,7 +407,7 @@ void main() {
           await tester.tap(hoverModeCheckBox);
           await tester.pump(safePumpDuration);
           expect(
-            preferences.inspectorV2.hoverEvalModeEnabled.value,
+            preferences.inspector.hoverEvalModeEnabled.value,
             !startingHoverEvalModeValue,
           );
         },